Add-Hyperlinks with more specific regex with the "section"(?) symbol
Several of the regex matches were wrong, like

    [Note: blah blah  See §[I.8.2.3](#todo-

So the find-replace was using the first `[` instead of the second `[` after the section symbol `§`.
Change the regex to hardcode the start `§[` which fixes many more links.

ALSO, manually fixed `ii.22.26-methoddef-0x06.md` which had the plural `hyperlinks` which didn't match the regex